Re-factored artifacts.
