+++ demeter-base/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>demeter-base</name>
+	<name>demeter2-base</name>
 	<comment></comment>
 	<projects>
 	</projects>
