+++ demeter-base/pom.xml
@@ -2,59 +2,50 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
-		<artifactId>demeter</artifactId>
+		<artifactId>demeter2</artifactId>
 		<groupId>com.heliosphere.demeter</groupId>
 		<version>0.0.1-SNAPSHOT</version>
 		<relativePath>../demeter</relativePath>
 	</parent>
-	<artifactId>demeter-base</artifactId>
+	<artifactId>demeter2-base</artifactId>
 	<name>Demeter Base</name>
 	<inceptionYear>2016</inceptionYear>
 	<dependencies>
 		<dependency>
 			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.11</version><!--$NO-MVN-MAN-VER$ -->
+			<artifactId>junit</artifactId><!--$NO-MVN-MAN-VER$ -->
 		</dependency>
 		<dependency>
 			<groupId>com.carrotsearch</groupId>
-			<artifactId>junit-benchmarks</artifactId>
-			<version>0.6.0</version><!--$NO-MVN-MAN-VER$ -->
+			<artifactId>junit-benchmarks</artifactId><!--$NO-MVN-MAN-VER$ -->
 		</dependency>
 		<dependency>
 			<groupId>log4j</groupId>
-			<artifactId>log4j</artifactId>
-			<version>1.2.14</version><!--$NO-MVN-MAN-VER$ -->
+			<artifactId>log4j</artifactId><!--$NO-MVN-MAN-VER$ -->
 		</dependency>
 		<dependency>
 			<groupId>commons-lang</groupId>
-			<artifactId>commons-lang</artifactId>
-			<version>2.6</version><!--$NO-MVN-MAN-VER$ -->
+			<artifactId>commons-lang</artifactId><!--$NO-MVN-MAN-VER$ -->
 		</dependency>
 		<dependency>
 			<groupId>commons-configuration</groupId>
-			<artifactId>commons-configuration</artifactId>
-			<version>1.9</version><!--$NO-MVN-MAN-VER$ -->
+			<artifactId>commons-configuration</artifactId><!--$NO-MVN-MAN-VER$ -->
 		</dependency>
 		<dependency>
 			<groupId>com.h2database</groupId>
-			<artifactId>h2</artifactId>
-			<version>1.3.160</version><!--$NO-MVN-MAN-VER$ -->
+			<artifactId>h2</artifactId><!--$NO-MVN-MAN-VER$ -->
 		</dependency>
 		<dependency>
 			<groupId>org.projectlombok</groupId>
-			<artifactId>lombok</artifactId>
-			<version>1.16.6</version><!--$NO-MVN-MAN-VER$ -->
+			<artifactId>lombok</artifactId><!--$NO-MVN-MAN-VER$ -->
 		</dependency>
 		<dependency>
 			<groupId>joda-time</groupId>
-			<artifactId>joda-time</artifactId>
-			<version>1.6</version><!--$NO-MVN-MAN-VER$ -->
+			<artifactId>joda-time</artifactId><!--$NO-MVN-MAN-VER$ -->
 		</dependency>
 		<dependency>
 			<groupId>commons-collections</groupId>
-			<artifactId>commons-collections</artifactId>
-			<version>3.2.1</version><!--$NO-MVN-MAN-VER$-->
+			<artifactId>commons-collections</artifactId><!--$NO-MVN-MAN-VER$-->
 		</dependency>
 		<dependency>
 			<groupId>com.google.guava</groupId>
