+++ demeter-base/run/junit/demeter-base - ResourceBundleTest.launch
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <launchConfiguration type="org.eclipse.jdt.junit.launchconfig">
 <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
-<listEntry value="/demeter-base/src/test/java/com/heliosphere/demeter/base/ResourceBundleTest.java"/>
+<listEntry value="/demeter2-base"/>
 </listAttribute>
 <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
-<listEntry value="1"/>
+<listEntry value="4"/>
 </listAttribute>
 <listAttribute key="org.eclipse.debug.ui.favoriteGroups">
 <listEntry value="org.eclipse.debug.ui.launchGroup.run"/>
@@ -15,7 +15,7 @@
 <stringAttribute key="org.eclipse.jdt.junit.TEST_KIND" value="org.eclipse.jdt.junit.loader.junit4"/>
 <booleanAttribute key="org.eclipse.jdt.launching.ATTR_USE_START_ON_FIRST_THREAD" value="true"/>
 <stringAttribute key="org.eclipse.jdt.launching.CLASSPATH_PROVIDER" value="org.eclipse.m2e.launchconfig.classpathProvider"/>
-<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="com.heliosphere.demeter.base.ResourceBundleTest"/>
-<stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="demeter-base"/>
+<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="com.heliosphere.demeter2.base.ResourceBundleTest"/>
+<stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="demeter2-base"/>
 <stringAttribute key="org.eclipse.jdt.launching.SOURCE_PATH_PROVIDER" value="org.eclipse.m2e.launchconfig.sourcepathProvider"/>
 </launchConfiguration>
