+++ demeter-base/src/main/java/com/heliosphere/demeter2/base/annotation/BundleEnum.java
@@ -9,7 +9,7 @@
  * License can be consulted at http://www.apache.org/licenses/LICENSE-2.0
  * ---------------------------------------------------------------------------
  */
-package com.heliosphere.demeter.base.annotation;
+package com.heliosphere.demeter2.base.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
