+++ demeter-base/src/main/java/com/heliosphere/demeter2/base/exception/InvalidArgumentException.java
@@ -9,17 +9,18 @@
  * License can be consulted at http://www.apache.org/licenses/LICENSE-2.0
  * ---------------------------------------------------------------------------
  */
-package com.heliosphere.demeter.base.exception;
+package com.heliosphere.demeter2.base.exception;
 
-import com.heliosphere.demeter.base.resource.bundle.IBundle;
+import com.heliosphere.demeter2.base.resource.bundle.IBundle;
 
 /**
- * Exception thrown to indicate that a service has not been implemented yet.
+ * Exception thrown to indicate a method has been passed an illegal or
+ * inappropriate argument.
  * <hr>
  * @author <a href="mailto:christophe.resse@gmail.com">Christophe Resse - Heliosphere</a>
  * @version 1.0.0
  */
-public class NotImplementedException extends AbstractUncheckedException
+public class InvalidArgumentException extends AbstractUncheckedException
 {
 	/**
 	 * Default serialization identifier.
@@ -27,27 +28,30 @@ public class NotImplementedException extends AbstractUncheckedException
 	private static final long serialVersionUID = 1L;
 
 	/**
-	 * Thrown to indicate that a service is not yet implemented.
+	 * Thrown to indicate that a method has been passed an illegal or
+	 * inappropriate argument.
 	 */
-	public NotImplementedException()
+	public InvalidArgumentException()
 	{
 		super();
 	}
 
 	/**
-	 * Thrown to indicate that a service is not yet implemented.
+	 * Thrown to indicate that a method has been passed an illegal or
+	 * inappropriate argument.
 	 * <p>
 	 * @param key Resource bundle key (enumerated value coming from an
 	 * enumeration implementing the {@link IBundle} interface).
 	 */
 	@SuppressWarnings("hiding")
-	public NotImplementedException(final Enum<? extends IBundle> key)
+	public InvalidArgumentException(final Enum<? extends IBundle> key)
 	{
 		super(key);
 	}
 
 	/**
-	 * Thrown to indicate that a service is not yet implemented.
+	 * Thrown to indicate that a method has been passed an illegal or
+	 * inappropriate argument.
 	 * <p>
 	 * @param key Exception key (enumerated value coming from an enumeration
 	 * implementing the {@link IExceptionType} interface).
@@ -55,40 +59,43 @@ public class NotImplementedException extends AbstractUncheckedException
 	 * message.
 	 */
 	@SuppressWarnings("hiding")
-	public NotImplementedException(final Enum<?> key, final Object... parameters)
+	public InvalidArgumentException(final Enum<?> key, final Object... parameters)
 	{
 		super(key, parameters);
 	}
 
 	/**
-	 * Thrown to indicate that a service is not yet implemented.
+	 * Thrown to indicate that a method has been passed an illegal or
+	 * inappropriate argument.
 	 * <p>
 	 * @param exception Parent {@link Exception}.
 	 */
-	public NotImplementedException(final Exception exception)
+	public InvalidArgumentException(final Exception exception)
 	{
 		super(exception);
 	}
 
 	/**
-	 * Thrown to indicate that a service is not yet implemented.
+	 * Thrown to indicate that a method has been passed an illegal or
+	 * inappropriate argument.
 	 * <p>
 	 * @param message Message describing the error being the cause of the raised
 	 * exception.
 	 */
-	public NotImplementedException(final String message)
+	public InvalidArgumentException(final String message)
 	{
 		super(message);
 	}
 
 	/**
-	 * Thrown to indicate that a service is not yet implemented.
+	 * Thrown to indicate that a method has been passed an illegal or
+	 * inappropriate argument.
 	 * <p>
 	 * @param message Message describing the error being the cause of the raised
 	 * exception.
 	 * @param exception Parent {@link Exception}.
 	 */
-	public NotImplementedException(final String message, final Exception exception)
+	public InvalidArgumentException(final String message, final Exception exception)
 	{
 		super(message, exception);
 	}
