+++ demeter-base/src/main/java/com/heliosphere/demeter2/base/resource/bundle/BundleDemeterBase.java
@@ -9,9 +9,9 @@
  * License can be consulted at http://www.apache.org/licenses/LICENSE-2.0
  * ---------------------------------------------------------------------------
  */
-package com.heliosphere.demeter.base.resource.bundle;
+package com.heliosphere.demeter2.base.resource.bundle;
 
-import com.heliosphere.demeter.base.annotation.BundleEnumRegister;
+import com.heliosphere.demeter2.base.annotation.BundleEnumRegister;
 
 /**
  * Enumeration of the resource bundles of the {@code demeter-base} module. Each
