+++ demeter/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>demeter</name>
+	<name>demeter2</name>
 	<comment></comment>
 	<projects>
 	</projects>
